Interpret `Pre-Depends`
 `Pre-Depends` field is more restrictive dependency than `Depends`.
 Pre-depended-on packages must be both extracted and configured before
Pre-depending packages are extracted.
(though normal depending-on packages are allowed to be extracted
before depended-on packages are configured.)

 To respect `Pre-Depends` fields, `rapt2` decides the order of extraction
and configuration of packages in below steps:

1. Regard `Pre-Depends` as normal `Depends` and do SCC and
topological sort.
2. Split sorted packages into layers. Each layer must start from a
pre-depended-on package if the layer contains pre-depended-on package.
3. For each layers, extract all packages in the layer, and configure
them afterward.

This strategy would NOT be robust.
However at least, this commit makes `rapt2` able to install
`python3`, which has circular dependency and `Pre-Depends`.